Refactoring code and cleaning up
