+++ functions.js
@@ -1,0 +1,4 @@
+const funcs = [];
+
+
+module.exports = funcs;
